Clean up wallet commands
