+++ cli/src/lib.rs
@@ -37,3 +37,4 @@ pub mod stake;
 pub mod test_utils;
 pub mod validator_info;
 pub mod vote;
+pub mod wallet;
